+++ src/om_inputs/schemas.cljs
@@ -46,6 +46,7 @@
    (s/optional-key :order) [sch-field]
    (s/optional-key :init) {sch-field s/Any}
    (s/optional-key :validations) s/Any
+   (s/optional-key :validate-i18n-keys) s/Bool
    (s/optional-key :action) {(s/optional-key :one-shot) s/Bool
                              (s/optional-key :no-reset) s/Bool
                              (s/optional-key :async) s/Bool